+++ bin/configs/swift-alt-petstore-new.yaml
@@ -1,5 +1,5 @@
 generatorName: swift-alt
-outputDir: samples/client/petstore/swift/alt
+outputDir: samples/client/petstore/swift-alt
 inputSpec: modules/openapi-generator/src/test/resources/3_0/petstore.yaml
 templateDir: modules/openapi-generator/src/main/resources/swift-alt
 additionalProperties:
