--- modules/openapi-generator/src/main/resources/swift-alt/Extensions.mustache
@@ -1,238 +1,0 @@
-// Extensions.swift
-//
-// Generated by openapi-generator
-// https://openapi-generator.tech
-//
-
-import Foundation
-#if canImport(AnyCodable)
-import AnyCodable
-#endif{{#usePromiseKit}}
-import PromiseKit{{/usePromiseKit}}{{#useVapor}}
-import Vapor{{/useVapor}}{{^useVapor}}
-
-extension Bool: JSONEncodable {
-    func encodeToJSON() -> Any { return self as Any }
-}
-
-extension Float: JSONEncodable {
-    func encodeToJSON() -> Any { return self as Any }
-}
-
-extension Int: JSONEncodable {
-    func encodeToJSON() -> Any { return self as Any }
-}
-
-extension Int32: JSONEncodable {
-    func encodeToJSON() -> Any { return NSNumber(value: self as Int32) }
-}
-
-extension Int64: JSONEncodable {
-    func encodeToJSON() -> Any { return NSNumber(value: self as Int64) }
-}
-
-extension Double: JSONEncodable {
-    func encodeToJSON() -> Any { return self as Any }
-}
-
-extension String: JSONEncodable {
-    func encodeToJSON() -> Any { return self as Any }
-}
-
-extension RawRepresentable where RawValue: JSONEncodable {
-    func encodeToJSON() -> Any { return self.rawValue as Any }
-}
-
-private func encodeIfPossible<T>(_ object: T) -> Any {
-    if let encodableObject = object as? JSONEncodable {
-        return encodableObject.encodeToJSON()
-    } else {
-        return object as Any
-    }
-}
-
-extension Array: JSONEncodable {
-    func encodeToJSON() -> Any {
-        return self.map(encodeIfPossible)
-    }
-}
-
-extension Set: JSONEncodable {
-    func encodeToJSON() -> Any {
-        return Array(self).encodeToJSON()
-    }
-}
-
-extension Dictionary: JSONEncodable {
-    func encodeToJSON() -> Any {
-        var dictionary = [AnyHashable: Any]()
-        for (key, value) in self {
-            dictionary[key] = encodeIfPossible(value)
-        }
-        return dictionary as Any
-    }
-}
-
-extension Data: JSONEncodable {
-    func encodeToJSON() -> Any {
-        return self.base64EncodedString(options: Data.Base64EncodingOptions())
-    }
-}
-
-extension Date: JSONEncodable {
-    func encodeToJSON() -> Any {
-        return CodableHelper.dateFormatter.string(from: self) as Any
-    }
-}
-
-extension URL: JSONEncodable {
-    func encodeToJSON() -> Any {
-        return self
-    }
-}
-
-extension UUID: JSONEncodable {
-    func encodeToJSON() -> Any {
-        return self.uuidString
-    }
-}{{/useVapor}}{{#generateModelAdditionalProperties}}
-
-extension String: CodingKey {
-
-    public var stringValue: String {
-        return self
-    }
-
-    public init?(stringValue: String) {
-        self.init(stringLiteral: stringValue)
-    }
-
-    public var intValue: Int? {
-        return nil
-    }
-
-    public init?(intValue: Int) {
-        return nil
-    }
-
-}
-
-extension KeyedEncodingContainerProtocol {
-
-    {{#nonPublicApi}}internal{{/nonPublicApi}}{{^nonPublicApi}}public{{/nonPublicApi}} mutating func encodeArray<T>(_ values: [T], forKey key: Self.Key) throws where T: Encodable {
-        var arrayContainer = nestedUnkeyedContainer(forKey: key)
-        try arrayContainer.encode(contentsOf: values)
-    }
-
-    {{#nonPublicApi}}internal{{/nonPublicApi}}{{^nonPublicApi}}public{{/nonPublicApi}} mutating func encodeArrayIfPresent<T>(_ values: [T]?, forKey key: Self.Key) throws where T: Encodable {
-        if let values = values {
-            try encodeArray(values, forKey: key)
-        }
-    }
-
-    {{#nonPublicApi}}internal{{/nonPublicApi}}{{^nonPublicApi}}public{{/nonPublicApi}} mutating func encodeMap<T>(_ pairs: [Self.Key: T]) throws where T: Encodable {
-        for (key, value) in pairs {
-            try encode(value, forKey: key)
-        }
-    }
-
-    {{#nonPublicApi}}internal{{/nonPublicApi}}{{^nonPublicApi}}public{{/nonPublicApi}} mutating func encodeMapIfPresent<T>(_ pairs: [Self.Key: T]?) throws where T: Encodable {
-        if let pairs = pairs {
-            try encodeMap(pairs)
-        }
-    }
-
-}
-
-extension KeyedDecodingContainerProtocol {
-
-    {{#nonPublicApi}}internal{{/nonPublicApi}}{{^nonPublicApi}}public{{/nonPublicApi}} func decodeArray<T>(_ type: T.Type, forKey key: Self.Key) throws -> [T] where T: Decodable {
-        var tmpArray = [T]()
-
-        var nestedContainer = try nestedUnkeyedContainer(forKey: key)
-        while !nestedContainer.isAtEnd {
-            let arrayValue = try nestedContainer.decode(T.self)
-            tmpArray.append(arrayValue)
-        }
-
-        return tmpArray
-    }
-
-    {{#nonPublicApi}}internal{{/nonPublicApi}}{{^nonPublicApi}}public{{/nonPublicApi}} func decodeArrayIfPresent<T>(_ type: T.Type, forKey key: Self.Key) throws -> [T]? where T: Decodable {
-        var tmpArray: [T]?
-
-        if contains(key) {
-            tmpArray = try decodeArray(T.self, forKey: key)
-        }
-
-        return tmpArray
-    }
-
-    {{#nonPublicApi}}internal{{/nonPublicApi}}{{^nonPublicApi}}public{{/nonPublicApi}} func decodeMap<T>(_ type: T.Type, excludedKeys: Set<Self.Key>) throws -> [Self.Key: T] where T: Decodable {
-        var map: [Self.Key: T] = [:]
-
-        for key in allKeys {
-            if !excludedKeys.contains(key) {
-                let value = try decode(T.self, forKey: key)
-                map[key] = value
-            }
-        }
-
-        return map
-    }
-
-}{{/generateModelAdditionalProperties}}{{^useVapor}}
-
-extension HTTPURLResponse {
-    var isStatusCodeSuccessful: Bool {
-        return Array(200 ..< 300).contains(statusCode)
-    }
-}{{/useVapor}}{{#usePromiseKit}}
-
-extension RequestBuilder {
-    {{#nonPublicApi}}internal{{/nonPublicApi}}{{^nonPublicApi}}public{{/nonPublicApi}} func execute() -> Promise<Response<T>> {
-        let deferred = Promise<Response<T>>.pending()
-        self.execute { result in
-            switch result {
-            case let .success(response):
-                deferred.resolver.fulfill(response)
-            case let .failure(error):
-                deferred.resolver.reject(error)
-            }
-        }
-        return deferred.promise
-    }
-}{{/usePromiseKit}}{{#useVapor}}
-
-extension UUID: Content { }
-
-extension URL: Content { }
-
-extension Bool: Content { }
-
-extension Set: ResponseEncodable where Element: Content {
-    public func encodeResponse(for request: Vapor.Request) -> EventLoopFuture<Vapor.Response> {
-        let response = Vapor.Response()
-        do {
-            try response.content.encode(Array(self))
-        } catch {
-            return request.eventLoop.makeFailedFuture(error)
-        }
-        return request.eventLoop.makeSucceededFuture(response)
-    }
-}
-
-extension Set: RequestDecodable where Element: Content {
-    public static func decodeRequest(_ request: Vapor.Request) -> EventLoopFuture<Self> {
-        do {
-            let content = try request.content.decode([Element].self)
-            return request.eventLoop.makeSucceededFuture(Set(content))
-        } catch {
-            return request.eventLoop.makeFailedFuture(error)
-        }
-    }
-}
-
-extension Set: Content where Element: Content { }
-
-extension AnyCodable: Content {}{{/useVapor}}
