--- modules/openapi-generator/src/main/resources/swift-alt/JSONEncodingHelper.mustache
@@ -1,45 +1,0 @@
-//
-// JSONEncodingHelper.swift
-//
-// Generated by openapi-generator
-// https://openapi-generator.tech
-//
-
-import Foundation
-
-{{#nonPublicApi}}internal{{/nonPublicApi}}{{^nonPublicApi}}open{{/nonPublicApi}} class JSONEncodingHelper {
-
-    {{#nonPublicApi}}internal{{/nonPublicApi}}{{^nonPublicApi}}open{{/nonPublicApi}} class func encodingParameters<T: Encodable>(forEncodableObject encodableObj: T?) -> [String: Any]? {
-        var params: [String: Any]?
-
-        // Encode the Encodable object
-        if let encodableObj = encodableObj {
-            let encodeResult = CodableHelper.encode(encodableObj)
-            do {
-                let data = try encodeResult.get()
-                params = JSONDataEncoding.encodingParameters(jsonData: data)
-            } catch {
-                print(error.localizedDescription)
-            }
-        }
-
-        return params
-    }
-
-    {{#nonPublicApi}}internal{{/nonPublicApi}}{{^nonPublicApi}}open{{/nonPublicApi}} class func encodingParameters(forEncodableObject encodableObj: Any?) -> [String: Any]? {
-        var params: [String: Any]?
-
-        if let encodableObj = encodableObj {
-            do {
-                let data = try JSONSerialization.data(withJSONObject: encodableObj, options: .prettyPrinted)
-                params = JSONDataEncoding.encodingParameters(jsonData: data)
-            } catch {
-                print(error.localizedDescription)
-                return nil
-            }
-        }
-
-        return params
-    }
-
-}
