--- modules/openapi-generator/src/main/resources/swift-alt/Models.mustache
@@ -1,54 +1,0 @@
-// Models.swift
-//
-// Generated by openapi-generator
-// https://openapi-generator.tech
-//
-
-import Foundation
-
-protocol JSONEncodable {
-    func encodeToJSON() -> Any
-}
-
-{{#nonPublicApi}}internal{{/nonPublicApi}}{{^nonPublicApi}}public{{/nonPublicApi}} enum ErrorResponse: Error {
-    case error(Int, Data?, URLResponse?, Error)
-}
-
-{{#nonPublicApi}}internal{{/nonPublicApi}}{{^nonPublicApi}}public{{/nonPublicApi}} enum DownloadException: Error {
-    case responseDataMissing
-    case responseFailed
-    case requestMissing
-    case requestMissingPath
-    case requestMissingURL
-}
-
-{{#nonPublicApi}}internal{{/nonPublicApi}}{{^nonPublicApi}}public{{/nonPublicApi}} enum DecodableRequestBuilderError: Error {
-    case emptyDataResponse
-    case nilHTTPResponse
-    case unsuccessfulHTTPStatusCode
-    case jsonDecoding(DecodingError)
-    case generalError(Error)
-}
-
-{{#nonPublicApi}}internal{{/nonPublicApi}}{{^nonPublicApi}}open{{/nonPublicApi}} class Response<T> {
-    {{#nonPublicApi}}internal{{/nonPublicApi}}{{^nonPublicApi}}public{{/nonPublicApi}} let statusCode: Int
-    {{#nonPublicApi}}internal{{/nonPublicApi}}{{^nonPublicApi}}public{{/nonPublicApi}} let header: [String: String]
-    {{#nonPublicApi}}internal{{/nonPublicApi}}{{^nonPublicApi}}public{{/nonPublicApi}} let body: T?
-
-    {{#nonPublicApi}}internal{{/nonPublicApi}}{{^nonPublicApi}}public{{/nonPublicApi}} init(statusCode: Int, header: [String: String], body: T?) {
-        self.statusCode = statusCode
-        self.header = header
-        self.body = body
-    }
-
-    {{#nonPublicApi}}internal{{/nonPublicApi}}{{^nonPublicApi}}public{{/nonPublicApi}} convenience init(response: HTTPURLResponse, body: T?) {
-        let rawHeader = response.allHeaderFields
-        var header = [String: String]()
-        for (key, value) in rawHeader {
-            if let key = key.base as? String, let value = value as? String {
-                header[key] = value
-            }
-        }
-        self.init(statusCode: response.statusCode, header: header, body: body)
-    }
-}
