+++ modules/openapi-generator/src/main/resources/swift-alt/Runtime.mustache
@@ -1,3 +1,8 @@
+// {{classname}}.swift
+//
+// Generated by openapi-generator
+// https://openapi-generator.tech
+
 import Foundation
 import Combine
