--- modules/openapi-generator/src/main/resources/swift-alt/XcodeGen.mustache
@@ -1,18 +1,0 @@
-name: {{projectName}}
-targets:
-  {{projectName}}:
-    type: framework
-    platform: iOS
-    deploymentTarget: "9.0"
-    sources: [{{swiftPackagePath}}{{^swiftPackagePath}}{{#useSPMFileStructure}}Sources{{/useSPMFileStructure}}{{^useSPMFileStructure}}{{projectName}}{{/useSPMFileStructure}}{{/swiftPackagePath}}]
-    info:
-      path: ./Info.plist
-      version: {{podVersion}}{{^podVersion}}{{#apiInfo}}{{version}}{{/apiInfo}}{{^apiInfo}}}0.0.1{{/apiInfo}}{{/podVersion}}
-    settings:
-      APPLICATION_EXTENSION_API_ONLY: true
-    scheme: {}
-    dependencies:
-    - carthage: AnyCodable{{#useAlamofire}}
-    - carthage: Alamofire{{/useAlamofire}}{{#usePromiseKit}}
-    - carthage: PromiseKit{{/usePromiseKit}}{{#useRxSwift}}
-    - carthage: RxSwift{{/useRxSwift}}
