--- modules/openapi-generator/src/main/resources/swift-alt/_param.mustache
@@ -1,1 +1,0 @@
-"{{baseName}}": {{paramName}}{{^required}}?{{/required}}.encodeToJSON()
