+++ modules/openapi-generator/src/main/resources/swift-alt/model.mustache
@@ -6,9 +6,6 @@
 //
 
 import Foundation
-#if canImport(AnyCodable)
-import AnyCodable
-#endif
 
 {{#description}}/** {{.}} */{{/description}}
 {{#vendorExtensions.x-is-one-of-interface}}
