+++ samples/client/petstore/swift-alt/.openapi-generator-ignore
