+++ samples/client/petstore/swift-alt/.swiftpm/xcode/package.xcworkspace/contents.xcworkspacedata
