+++ samples/client/petstore/swift-alt/Package.swift
@@ -1,0 +1,24 @@
+// swift-tools-version:5.1
+
+import PackageDescription
+
+let package = Package(
+    name: "OpenAPIClient",
+    platforms: [
+        .iOS(.v13)
+    ],
+    products: [
+        .library(
+            name: "OpenAPIClient",
+            targets: ["OpenAPIClient"]
+        ),
+    ],
+    dependencies: [],
+    targets: [
+        .target(
+            name: "OpenAPIClient",
+            dependencies: [],
+            path: "Sources"
+        ),
+    ]
+)
