+++ samples/client/petstore/swift-alt/Sources/Models/ApiResponse.swift
@@ -1,0 +1,37 @@
+//
+// ApiResponse.swift
+//
+// Generated by openapi-generator
+// https://openapi-generator.tech
+//
+
+import Foundation
+
+/** Describes the result of uploading an image resource */
+
+public struct ApiResponse: Codable, Hashable {
+    public var code: Int?
+    public var type: String?
+    public var message: String?
+
+    public init(code: Int? = nil, type: String? = nil, message: String? = nil) {
+        self.code = code
+        self.type = type
+        self.message = message
+    }
+
+    public enum CodingKeys: String, CodingKey, CaseIterable {
+        case code
+        case type
+        case message
+    }
+
+    // Encodable protocol methods
+
+    public func encode(to encoder: Encoder) throws {
+        var container = encoder.container(keyedBy: CodingKeys.self)
+        try container.encodeIfPresent(code, forKey: .code)
+        try container.encodeIfPresent(type, forKey: .type)
+        try container.encodeIfPresent(message, forKey: .message)
+    }
+}
