+++ samples/client/petstore/swift-alt/Sources/Runtime.swift
@@ -1,3 +1,8 @@
+// .swift
+//
+// Generated by openapi-generator
+// https://openapi-generator.tech
+
 import Foundation
 import Combine
