--- samples/client/petstore/swift/alt/.openapi-generator/FILES
@@ -1,11 +1,0 @@
-OpenAPIClient/Classes/OpenAPIs/APIs/PetAPI.swift
-OpenAPIClient/Classes/OpenAPIs/APIs/StoreAPI.swift
-OpenAPIClient/Classes/OpenAPIs/APIs/UserAPI.swift
-OpenAPIClient/Classes/OpenAPIs/Models/ApiResponse.swift
-OpenAPIClient/Classes/OpenAPIs/Models/Category.swift
-OpenAPIClient/Classes/OpenAPIs/Models/Order.swift
-OpenAPIClient/Classes/OpenAPIs/Models/Pet.swift
-OpenAPIClient/Classes/OpenAPIs/Models/Tag.swift
-OpenAPIClient/Classes/OpenAPIs/Models/User.swift
-Package.swift
-project.yml
