--- samples/client/petstore/swift/alt/OpenAPIClient/Classes/OpenAPIs/APIs/PetAPI.swift
@@ -1,391 +1,0 @@
-//
-// PetAPI.swift
-//
-// Generated by openapi-generator
-// https://openapi-generator.tech
-// Run in root folder to build example mvn clean package -DskipTests; ./bin/generate-samples.sh bin/configs/swift-alt-petstore-new.yaml
-
-import Foundation
-import Combine
-// import Runtime
-
-
-open class PetAPI {
-    private let encoder: JSONEncoder = {
-        let encoder = JSONEncoder()
-        encoder.dateEncodingStrategy = .formatted(OpenISO8601DateFormatter())
-        encoder.outputFormatting = .prettyPrinted
-        return encoder
-    }()
-    private let decoder: JSONDecoder = {
-        let decoder = JSONDecoder()
-        decoder.dateDecodingStrategy = .formatted(OpenISO8601DateFormatter())
-        return decoder
-    }()
-    private let transport: Transport
-
-    public init(_ transport: Transport) {
-        self.transport = transport
-    }
-
-    /**
-        Add a new pet to the store
-        - POST /pet
-        - OAuth:
-        - type: oauth2
-        - name: petstore_auth
-        - parameter pet: (body) Pet object that needs to be added to the store 
-        - returns: AnyPublisher<Result<Pet, Error>, Never> 
-        */
-    open func addPet(pet: Pet) -> AnyPublisher<Result<Pet, Error>, Never> {
-        // Creating final URL with query items and path
-        var components = URLComponents()
-        components.path = "/pet"
-        
-        guard let url = components.url(relativeTo: transport.baseURL) else {
-            fatalError("URL is nil")
-        }
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "POST"
-        
-        // Setting body parameters 
-        request.httpBody = try? encoder.encode(pet)
-        // Getting auth type 
-        let securityScheme: SecurityScheme = .none
-        return transport.send(request: request, securityScheme: securityScheme)
-            .map { response -> Result<Pet, Error> in
-                if response.response?.statusCode == 200 {
-                    if let data = response.data {
-                        return Result { try self.decoder.decode(Pet.self, from: data) }
-                    } else {
-                        return .failure(TransportError.unknown)
-                    }
-                } else {
-                    return .failure(response.error ?? TransportError.unknown)
-                }
-            }
-            .eraseToAnyPublisher()
-    }
-
-    /**
-        Deletes a pet
-        - DELETE /pet/{petId}
-        - OAuth:
-        - type: oauth2
-        - name: petstore_auth
-        - parameter petId: (path) Pet id to delete 
-        - parameter apiKey: (header)  (optional)
-        - returns: AnyPublisher<Result<Void, Error>, Never> 
-        */
-    open func deletePet(petId: Int64, apiKey: String? = nil) -> AnyPublisher<Result<Void, Error>, Never> {
-        // Creating final URL with query items and path
-        var components = URLComponents()
-        components.path = "/pet/{petId}"
-            .replacingOccurrences(of: "{petId}", with: "\(petId)")
-        
-        guard let url = components.url(relativeTo: transport.baseURL) else {
-            fatalError("URL is nil")
-        }
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "DELETE"
-        // Setting headers
-        request.allHTTPHeaderFields = [
-            "api_key": apiKey
-        ].compactMapValues { $0 }
-        
-        // Getting auth type 
-        let securityScheme: SecurityScheme = .none
-        return transport.send(request: request, securityScheme: securityScheme)
-            .map { response -> Result<Void, Error> in
-                if response.response?.statusCode == 200 {
-                    return .success(())
-                } else {
-                    return .failure(response.error ?? TransportError.unknown)
-                }
-            }
-            .eraseToAnyPublisher()
-    }
-
-    /**
-        Finds Pets by status
-        - GET /pet/findByStatus
-        - Multiple status values can be provided with comma separated strings
-        - OAuth:
-        - type: oauth2
-        - name: petstore_auth
-        - parameter status: (query) Status values that need to be considered for filter 
-        - returns: AnyPublisher<Result<[Pet], Error>, Never> 
-        */
-    open func findPetsByStatus(status: [String]) -> AnyPublisher<Result<[Pet], Error>, Never> {
-        // Creating final URL with query items and path
-        var components = URLComponents()
-        components.path = "/pet/findByStatus"
-        components.queryItems = [
-            URLQueryItem(name: "status", value: status.joined(separator: ", "))
-        ]
-        guard let url = components.url(relativeTo: transport.baseURL) else {
-            fatalError("URL is nil")
-        }
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "GET"
-        
-        
-        // Getting auth type 
-        let securityScheme: SecurityScheme = .none
-        return transport.send(request: request, securityScheme: securityScheme)
-            .map { response -> Result<[Pet], Error> in
-                if response.response?.statusCode == 200 {
-                    if let data = response.data {
-                        return Result { try self.decoder.decode([Pet].self, from: data) }
-                    } else {
-                        return .failure(TransportError.unknown)
-                    }
-                } else {
-                    return .failure(response.error ?? TransportError.unknown)
-                }
-            }
-            .eraseToAnyPublisher()
-    }
-
-    /**
-        Finds Pets by tags
-        - GET /pet/findByTags
-        - Multiple tags can be provided with comma separated strings. Use tag1, tag2, tag3 for testing.
-        - OAuth:
-        - type: oauth2
-        - name: petstore_auth
-        - parameter tags: (query) Tags to filter by 
-        - returns: AnyPublisher<Result<[Pet], Error>, Never> 
-        */
-    open func findPetsByTags(tags: [String]) -> AnyPublisher<Result<[Pet], Error>, Never> {
-        // Creating final URL with query items and path
-        var components = URLComponents()
-        components.path = "/pet/findByTags"
-        components.queryItems = [
-            URLQueryItem(name: "tags", value: tags.joined(separator: ", "))
-        ]
-        guard let url = components.url(relativeTo: transport.baseURL) else {
-            fatalError("URL is nil")
-        }
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "GET"
-        
-        
-        // Getting auth type 
-        let securityScheme: SecurityScheme = .none
-        return transport.send(request: request, securityScheme: securityScheme)
-            .map { response -> Result<[Pet], Error> in
-                if response.response?.statusCode == 200 {
-                    if let data = response.data {
-                        return Result { try self.decoder.decode([Pet].self, from: data) }
-                    } else {
-                        return .failure(TransportError.unknown)
-                    }
-                } else {
-                    return .failure(response.error ?? TransportError.unknown)
-                }
-            }
-            .eraseToAnyPublisher()
-    }
-
-    /**
-        Find pet by ID
-        - GET /pet/{petId}
-        - Returns a single pet
-        - API Key:
-        - type: apiKey api_key 
-        - name: api_key
-        - parameter petId: (path) ID of pet to return 
-        - returns: AnyPublisher<Result<Pet, Error>, Never> 
-        */
-    open func getPetById(petId: Int64) -> AnyPublisher<Result<Pet, Error>, Never> {
-        // Creating final URL with query items and path
-        var components = URLComponents()
-        components.path = "/pet/{petId}"
-            .replacingOccurrences(of: "{petId}", with: "\(petId)")
-        
-        guard let url = components.url(relativeTo: transport.baseURL) else {
-            fatalError("URL is nil")
-        }
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "GET"
-        
-        
-        // Getting auth type 
-        let securityScheme: SecurityScheme = .none
-        return transport.send(request: request, securityScheme: securityScheme)
-            .map { response -> Result<Pet, Error> in
-                if response.response?.statusCode == 200 {
-                    if let data = response.data {
-                        return Result { try self.decoder.decode(Pet.self, from: data) }
-                    } else {
-                        return .failure(TransportError.unknown)
-                    }
-                } else {
-                    return .failure(response.error ?? TransportError.unknown)
-                }
-            }
-            .eraseToAnyPublisher()
-    }
-
-    /**
-        Update an existing pet
-        - PUT /pet
-        - OAuth:
-        - type: oauth2
-        - name: petstore_auth
-        - parameter pet: (body) Pet object that needs to be added to the store 
-        - returns: AnyPublisher<Result<Pet, Error>, Never> 
-        */
-    open func updatePet(pet: Pet) -> AnyPublisher<Result<Pet, Error>, Never> {
-        // Creating final URL with query items and path
-        var components = URLComponents()
-        components.path = "/pet"
-        
-        guard let url = components.url(relativeTo: transport.baseURL) else {
-            fatalError("URL is nil")
-        }
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "PUT"
-        
-        // Setting body parameters 
-        request.httpBody = try? encoder.encode(pet)
-        // Getting auth type 
-        let securityScheme: SecurityScheme = .none
-        return transport.send(request: request, securityScheme: securityScheme)
-            .map { response -> Result<Pet, Error> in
-                if response.response?.statusCode == 200 {
-                    if let data = response.data {
-                        return Result { try self.decoder.decode(Pet.self, from: data) }
-                    } else {
-                        return .failure(TransportError.unknown)
-                    }
-                } else {
-                    return .failure(response.error ?? TransportError.unknown)
-                }
-            }
-            .eraseToAnyPublisher()
-    }
-
-    /**
-        Updates a pet in the store with form data
-        - POST /pet/{petId}
-        - OAuth:
-        - type: oauth2
-        - name: petstore_auth
-        - parameter petId: (path) ID of pet that needs to be updated 
-        - parameter name: (form) Updated name of the pet (optional)
-        - parameter status: (form) Updated status of the pet (optional)
-        - returns: AnyPublisher<Result<Void, Error>, Never> 
-        */
-    open func updatePetWithForm(petId: Int64, name: String? = nil, status: String? = nil) -> AnyPublisher<Result<Void, Error>, Never> {
-        // Creating final URL with query items and path
-        var components = URLComponents()
-        components.path = "/pet/{petId}"
-            .replacingOccurrences(of: "{petId}", with: "\(petId)")
-        
-        guard let url = components.url(relativeTo: transport.baseURL) else {
-            fatalError("URL is nil")
-        }
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "POST"
-        
-        
-        // Getting auth type 
-        let securityScheme: SecurityScheme = .none
-        return transport.send(request: request, securityScheme: securityScheme)
-            .map { response -> Result<Void, Error> in
-                if response.response?.statusCode == 200 {
-                    return .success(())
-                } else {
-                    return .failure(response.error ?? TransportError.unknown)
-                }
-            }
-            .eraseToAnyPublisher()
-    }
-
-    /**
-        uploads an image
-        - POST /pet/{petId}/uploadImage
-        - OAuth:
-        - type: oauth2
-        - name: petstore_auth
-        - parameter petId: (path) ID of pet to update 
-        - parameter additionalMetadata: (form) Additional data to pass to server (optional)
-        - parameter file: (form) file to upload (optional)
-        - returns: AnyPublisher<Result<ApiResponse, Error>, Never> 
-        */
-    open func uploadFile(petId: Int64, additionalMetadata: String? = nil, file: Data? = nil) -> AnyPublisher<Result<ApiResponse, Error>, Never> {
-        // Creating final URL with query items and path
-        var components = URLComponents()
-        components.path = "/pet/{petId}/uploadImage"
-            .replacingOccurrences(of: "{petId}", with: "\(petId)")
-        
-        guard let url = components.url(relativeTo: transport.baseURL) else {
-            fatalError("URL is nil")
-        }
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "POST"
-        
-        
-        // Getting auth type 
-        let securityScheme: SecurityScheme = .none
-        return transport.send(request: request, securityScheme: securityScheme)
-            .map { response -> Result<ApiResponse, Error> in
-                if response.response?.statusCode == 200 {
-                    if let data = response.data {
-                        return Result { try self.decoder.decode(ApiResponse.self, from: data) }
-                    } else {
-                        return .failure(TransportError.unknown)
-                    }
-                } else {
-                    return .failure(response.error ?? TransportError.unknown)
-                }
-            }
-            .eraseToAnyPublisher()
-    }
-}
-
-
-private class OpenISO8601DateFormatter: DateFormatter {
-    static let withoutSeconds: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.calendar = Calendar(identifier: .iso8601)
-        formatter.locale = Locale(identifier: "en_US_POSIX")
-        formatter.timeZone = TimeZone(secondsFromGMT: 0)
-        formatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ssZZZZZ"
-        return formatter
-    }()
-
-    private func setup() {
-        calendar = Calendar(identifier: .iso8601)
-        locale = Locale(identifier: "en_US_POSIX")
-        timeZone = TimeZone(secondsFromGMT: 0)
-        dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSSZZZZZ"
-    }
-
-    override init() {
-        super.init()
-        setup()
-    }
-
-    required init?(coder aDecoder: NSCoder) {
-        super.init(coder: aDecoder)
-        setup()
-    }
-
-    override func date(from string: String) -> Date? {
-        if let result = super.date(from: string) {
-            return result
-        }
-        return OpenISO8601DateFormatter.withoutSeconds.date(from: string)
-    }
-}
