--- samples/client/petstore/swift/alt/OpenAPIClient/Classes/OpenAPIs/APIs/UserAPI.swift
@@ -1,363 +1,0 @@
-//
-// UserAPI.swift
-//
-// Generated by openapi-generator
-// https://openapi-generator.tech
-// Run in root folder to build example mvn clean package -DskipTests; ./bin/generate-samples.sh bin/configs/swift-alt-petstore-new.yaml
-
-import Foundation
-import Combine
-// import Runtime
-
-
-open class UserAPI {
-    private let encoder: JSONEncoder = {
-        let encoder = JSONEncoder()
-        encoder.dateEncodingStrategy = .formatted(OpenISO8601DateFormatter())
-        encoder.outputFormatting = .prettyPrinted
-        return encoder
-    }()
-    private let decoder: JSONDecoder = {
-        let decoder = JSONDecoder()
-        decoder.dateDecodingStrategy = .formatted(OpenISO8601DateFormatter())
-        return decoder
-    }()
-    private let transport: Transport
-
-    public init(_ transport: Transport) {
-        self.transport = transport
-    }
-
-    /**
-        Create user
-        - POST /user
-        - This can only be done by the logged in user.
-        - API Key:
-        - type: apiKey api_key 
-        - name: api_key
-        - parameter user: (body) Created user object 
-        - returns: AnyPublisher<Result<Void, Error>, Never> 
-        */
-    open func createUser(user: User) -> AnyPublisher<Result<Void, Error>, Never> {
-        // Creating final URL with query items and path
-        var components = URLComponents()
-        components.path = "/user"
-        
-        guard let url = components.url(relativeTo: transport.baseURL) else {
-            fatalError("URL is nil")
-        }
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "POST"
-        
-        // Setting body parameters 
-        request.httpBody = try? encoder.encode(user)
-        // Getting auth type 
-        let securityScheme: SecurityScheme = .none
-        return transport.send(request: request, securityScheme: securityScheme)
-            .map { response -> Result<Void, Error> in
-                if response.response?.statusCode == 200 {
-                    return .success(())
-                } else {
-                    return .failure(response.error ?? TransportError.unknown)
-                }
-            }
-            .eraseToAnyPublisher()
-    }
-
-    /**
-        Creates list of users with given input array
-        - POST /user/createWithArray
-        - API Key:
-        - type: apiKey api_key 
-        - name: api_key
-        - parameter user: (body) List of user object 
-        - returns: AnyPublisher<Result<Void, Error>, Never> 
-        */
-    open func createUsersWithArrayInput(user: [User]) -> AnyPublisher<Result<Void, Error>, Never> {
-        // Creating final URL with query items and path
-        var components = URLComponents()
-        components.path = "/user/createWithArray"
-        
-        guard let url = components.url(relativeTo: transport.baseURL) else {
-            fatalError("URL is nil")
-        }
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "POST"
-        
-        // Setting body parameters 
-        request.httpBody = try? encoder.encode(user)
-        // Getting auth type 
-        let securityScheme: SecurityScheme = .none
-        return transport.send(request: request, securityScheme: securityScheme)
-            .map { response -> Result<Void, Error> in
-                if response.response?.statusCode == 200 {
-                    return .success(())
-                } else {
-                    return .failure(response.error ?? TransportError.unknown)
-                }
-            }
-            .eraseToAnyPublisher()
-    }
-
-    /**
-        Creates list of users with given input array
-        - POST /user/createWithList
-        - API Key:
-        - type: apiKey api_key 
-        - name: api_key
-        - parameter user: (body) List of user object 
-        - returns: AnyPublisher<Result<Void, Error>, Never> 
-        */
-    open func createUsersWithListInput(user: [User]) -> AnyPublisher<Result<Void, Error>, Never> {
-        // Creating final URL with query items and path
-        var components = URLComponents()
-        components.path = "/user/createWithList"
-        
-        guard let url = components.url(relativeTo: transport.baseURL) else {
-            fatalError("URL is nil")
-        }
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "POST"
-        
-        // Setting body parameters 
-        request.httpBody = try? encoder.encode(user)
-        // Getting auth type 
-        let securityScheme: SecurityScheme = .none
-        return transport.send(request: request, securityScheme: securityScheme)
-            .map { response -> Result<Void, Error> in
-                if response.response?.statusCode == 200 {
-                    return .success(())
-                } else {
-                    return .failure(response.error ?? TransportError.unknown)
-                }
-            }
-            .eraseToAnyPublisher()
-    }
-
-    /**
-        Delete user
-        - DELETE /user/{username}
-        - This can only be done by the logged in user.
-        - API Key:
-        - type: apiKey api_key 
-        - name: api_key
-        - parameter username: (path) The name that needs to be deleted 
-        - returns: AnyPublisher<Result<Void, Error>, Never> 
-        */
-    open func deleteUser(username: String) -> AnyPublisher<Result<Void, Error>, Never> {
-        // Creating final URL with query items and path
-        var components = URLComponents()
-        components.path = "/user/{username}"
-            .replacingOccurrences(of: "{username}", with: "\(username)")
-        
-        guard let url = components.url(relativeTo: transport.baseURL) else {
-            fatalError("URL is nil")
-        }
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "DELETE"
-        
-        
-        // Getting auth type 
-        let securityScheme: SecurityScheme = .none
-        return transport.send(request: request, securityScheme: securityScheme)
-            .map { response -> Result<Void, Error> in
-                if response.response?.statusCode == 200 {
-                    return .success(())
-                } else {
-                    return .failure(response.error ?? TransportError.unknown)
-                }
-            }
-            .eraseToAnyPublisher()
-    }
-
-    /**
-        Get user by user name
-        - GET /user/{username}
-        - parameter username: (path) The name that needs to be fetched. Use user1 for testing. 
-        - returns: AnyPublisher<Result<User, Error>, Never> 
-        */
-    open func getUserByName(username: String) -> AnyPublisher<Result<User, Error>, Never> {
-        // Creating final URL with query items and path
-        var components = URLComponents()
-        components.path = "/user/{username}"
-            .replacingOccurrences(of: "{username}", with: "\(username)")
-        
-        guard let url = components.url(relativeTo: transport.baseURL) else {
-            fatalError("URL is nil")
-        }
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "GET"
-        
-        
-        // Getting auth type 
-        let securityScheme: SecurityScheme = .none
-        return transport.send(request: request, securityScheme: securityScheme)
-            .map { response -> Result<User, Error> in
-                if response.response?.statusCode == 200 {
-                    if let data = response.data {
-                        return Result { try self.decoder.decode(User.self, from: data) }
-                    } else {
-                        return .failure(TransportError.unknown)
-                    }
-                } else {
-                    return .failure(response.error ?? TransportError.unknown)
-                }
-            }
-            .eraseToAnyPublisher()
-    }
-
-    /**
-        Logs user into the system
-        - GET /user/login
-        - responseHeaders: [Set-Cookie(String), X-Rate-Limit(Int), X-Expires-After(Date)]
-        - parameter username: (query) The user name for login 
-        - parameter password: (query) The password for login in clear text 
-        - returns: AnyPublisher<Result<String, Error>, Never> 
-        */
-    open func loginUser(username: String, password: String) -> AnyPublisher<Result<String, Error>, Never> {
-        // Creating final URL with query items and path
-        var components = URLComponents()
-        components.path = "/user/login"
-        components.queryItems = [
-            URLQueryItem(name: "username", value: username), 
-            URLQueryItem(name: "password", value: password)
-        ]
-        guard let url = components.url(relativeTo: transport.baseURL) else {
-            fatalError("URL is nil")
-        }
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "GET"
-        
-        
-        // Getting auth type 
-        let securityScheme: SecurityScheme = .none
-        return transport.send(request: request, securityScheme: securityScheme)
-            .map { response -> Result<String, Error> in
-                if response.response?.statusCode == 200 {
-                    if let data = response.data {
-                        return Result { try self.decoder.decode(String.self, from: data) }
-                    } else {
-                        return .failure(TransportError.unknown)
-                    }
-                } else {
-                    return .failure(response.error ?? TransportError.unknown)
-                }
-            }
-            .eraseToAnyPublisher()
-    }
-
-    /**
-        Logs out current logged in user session
-        - GET /user/logout
-        - API Key:
-        - type: apiKey api_key 
-        - name: api_key
-        - returns: AnyPublisher<Result<Void, Error>, Never> 
-        */
-    open func logoutUser() -> AnyPublisher<Result<Void, Error>, Never> {
-        // Creating final URL with query items and path
-        var components = URLComponents()
-        components.path = "/user/logout"
-        
-        guard let url = components.url(relativeTo: transport.baseURL) else {
-            fatalError("URL is nil")
-        }
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "GET"
-        
-        
-        // Getting auth type 
-        let securityScheme: SecurityScheme = .none
-        return transport.send(request: request, securityScheme: securityScheme)
-            .map { response -> Result<Void, Error> in
-                if response.response?.statusCode == 200 {
-                    return .success(())
-                } else {
-                    return .failure(response.error ?? TransportError.unknown)
-                }
-            }
-            .eraseToAnyPublisher()
-    }
-
-    /**
-        Updated user
-        - PUT /user/{username}
-        - This can only be done by the logged in user.
-        - API Key:
-        - type: apiKey api_key 
-        - name: api_key
-        - parameter username: (path) name that need to be deleted 
-        - parameter user: (body) Updated user object 
-        - returns: AnyPublisher<Result<Void, Error>, Never> 
-        */
-    open func updateUser(username: String, user: User) -> AnyPublisher<Result<Void, Error>, Never> {
-        // Creating final URL with query items and path
-        var components = URLComponents()
-        components.path = "/user/{username}"
-            .replacingOccurrences(of: "{username}", with: "\(username)")
-        
-        guard let url = components.url(relativeTo: transport.baseURL) else {
-            fatalError("URL is nil")
-        }
-
-        var request = URLRequest(url: url)
-        request.httpMethod = "PUT"
-        
-        // Setting body parameters 
-        request.httpBody = try? encoder.encode(user)
-        // Getting auth type 
-        let securityScheme: SecurityScheme = .none
-        return transport.send(request: request, securityScheme: securityScheme)
-            .map { response -> Result<Void, Error> in
-                if response.response?.statusCode == 200 {
-                    return .success(())
-                } else {
-                    return .failure(response.error ?? TransportError.unknown)
-                }
-            }
-            .eraseToAnyPublisher()
-    }
-}
-
-
-private class OpenISO8601DateFormatter: DateFormatter {
-    static let withoutSeconds: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.calendar = Calendar(identifier: .iso8601)
-        formatter.locale = Locale(identifier: "en_US_POSIX")
-        formatter.timeZone = TimeZone(secondsFromGMT: 0)
-        formatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ssZZZZZ"
-        return formatter
-    }()
-
-    private func setup() {
-        calendar = Calendar(identifier: .iso8601)
-        locale = Locale(identifier: "en_US_POSIX")
-        timeZone = TimeZone(secondsFromGMT: 0)
-        dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSSZZZZZ"
-    }
-
-    override init() {
-        super.init()
-        setup()
-    }
-
-    required init?(coder aDecoder: NSCoder) {
-        super.init(coder: aDecoder)
-        setup()
-    }
-
-    override func date(from string: String) -> Date? {
-        if let result = super.date(from: string) {
-            return result
-        }
-        return OpenISO8601DateFormatter.withoutSeconds.date(from: string)
-    }
-}
