--- samples/client/petstore/swift/alt/OpenAPIClient/Classes/OpenAPIs/Models/Tag.swift
@@ -1,36 +1,0 @@
-//
-// Tag.swift
-//
-// Generated by openapi-generator
-// https://openapi-generator.tech
-//
-
-import Foundation
-#if canImport(AnyCodable)
-import AnyCodable
-#endif
-
-/** A tag for a pet */
-
-public struct Tag: Codable, Hashable {
-    public var id: Int64?
-    public var name: String?
-
-    public init(id: Int64? = nil, name: String? = nil) {
-        self.id = id
-        self.name = name
-    }
-
-    public enum CodingKeys: String, CodingKey, CaseIterable {
-        case id
-        case name
-    }
-
-    // Encodable protocol methods
-
-    public func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        try container.encodeIfPresent(id, forKey: .id)
-        try container.encodeIfPresent(name, forKey: .name)
-    }
-}
