--- samples/client/petstore/swift/alt/project.yml
@@ -1,15 +1,0 @@
-name: OpenAPIClient
-targets:
-  OpenAPIClient:
-    type: framework
-    platform: iOS
-    deploymentTarget: "9.0"
-    sources: [OpenAPIClient]
-    info:
-      path: ./Info.plist
-      version: 1.0.0
-    settings:
-      APPLICATION_EXTENSION_API_ONLY: true
-    scheme: {}
-    dependencies:
-    - carthage: AnyCodable
